enforce float32

+++ models/cunumeric/grayscott_gc.jl
@@ -3,12 +3,12 @@ using Legate
 using Printf
 
 struct Params
-    dx::Float64
-    dt::Float64
-    c_u::Float64
-    c_v::Float64
-    f::Float64
-    k::Float64
+    dx::Float32
+    dt::Float32
+    c_u::Float32
+    c_v::Float32
+    f::Float32
+    k::Float32
 
     function Params(dx=1, c_u=1.0, c_v=0.3, f=0.03, k=0.06)
         new(dx, dx/5, c_u, c_v, f, k)
@@ -85,7 +85,7 @@ function grayscott(N, M, n_steps, gc_interval)
 
     u[1:150, 1:150] = cuNumeric.random(FT, (150, 150))
     v[1:150, 1:150] = cuNumeric.random(FT, (150, 150))
-
+    
     for n in 1:n_steps
         step(u, v, u_new, v_new, args)
         # update u and v
